Port Scanner using Socket
When we run the above script, it will prompt for the hostname, you can provide any hostname like the name of any website but be careful because port scanning can be seen as, or construed as, a crime. We should never execute a port scanner against any website or IP address without explicit, written permission from the owner of the server or computer that you are targeting. Port scanning is akin to going to someone’s house and checking their doors and windows. That is why it is advisable to use a port scanner on localhost or your own website (if any).